Changed DirectFile.cpp, DirectFile.h, FftRulerView.cpp
Added OpenAllowReadOnlyFallback flag for Open();
Fixed bug: bit for the last 64K in WrittenMask was not set.
Added source file support.
Prefetch request doesn't override active prefetch for the same range.
Prefetch request is limited to the size of the file.
Buffer is not discarded, if the returned range is not multiple of 64K;
Added time limit to FlushDirtyBuffers;
Added time limit to InitializeTheRestOfFile;
Added IsOpen(), IsReadOnly functions